requirements.txt
glom==22.1.0
pandas==1.4.3
progress==1.6
requests==2.28.0
websocket_client==1.3.3
